+++ src/models/reply.ts
@@ -1,11 +1,12 @@
 import { type DateTime } from 'luxon'
-import { BaseModel, column, belongsTo, hasMany, scope } from '@adonisjs/lucid/orm'
+import { column, belongsTo, hasMany, scope } from '@adonisjs/lucid/orm'
+import EscalatedBaseModel from './base_model.js'
 import type { BelongsTo, HasMany } from '@adonisjs/lucid/types/relations'
 import type { UserId } from '../helpers/user_id_column.js'
 import Ticket from './ticket.js'
 import Attachment from './attachment.js'
 
-export default class Reply extends BaseModel {
+export default class Reply extends EscalatedBaseModel {
   static table = 'escalated_replies'
 
   @column({ isPrimary: true })