+++ src/models/ticket_activity.ts
@@ -1,11 +1,12 @@
 import { type DateTime } from 'luxon'
-import { BaseModel, column, belongsTo, computed } from '@adonisjs/lucid/orm'
+import { column, belongsTo, computed } from '@adonisjs/lucid/orm'
+import EscalatedBaseModel from './base_model.js'
 import type { BelongsTo } from '@adonisjs/lucid/types/relations'
 import type { UserId } from '../helpers/user_id_column.js'
 import type { ActivityType } from '../types.js'
 import Ticket from './ticket.js'
 
-export default class TicketActivity extends BaseModel {
+export default class TicketActivity extends EscalatedBaseModel {
   static table = 'escalated_ticket_activities'
 
   @column({ isPrimary: true })